requirements.txt
aspose_cells==22.6.0
fake_useragent==0.1.11
JPype1==1.4.0
pyshadow==0.0.4
selenium==4.3.0
undetected_chromedriver==3.1.5.post4
webdriver_manager==3.7.0
